Show group icon if video has been shared
