+++ app/src/main/res/layout/item_browser_grid.xml
@@ -57,6 +57,15 @@
         <ImageButton
             android:layout_width="48dp"
             android:layout_height="48dp"
+            android:id="@+id/sharedButton"
+            android:src="@drawable/ic_group_white_24dp"
+            android:background="@android:color/transparent"
+            android:scaleType="center"
+            android:layout_margin="8dp" />
+
+        <ImageButton
+            android:layout_width="48dp"
+            android:layout_height="48dp"
             android:id="@+id/uploadButton"
             android:src="@drawable/ic_file_download_white_24dp"
             android:background="@android:color/transparent"
